requirements.txt
streamlit
opencv-python-headless
numpy
matplotlib
Pillow
imageio[ffmpeg]
scikit-image
